+++ packages/fhir-normalize/src/validate/index.ts
@@ -1,8 +1,3 @@
-export {
-  COMMON_ELEMENT,
-  VALIDATION_CODE,
-  VALIDATION_MESSAGE,
-  VALIDATION_SEVERITY,
-} from './constants';
+export { VALIDATION_CODE, VALIDATION_MESSAGE, VALIDATION_SEVERITY } from './constants';
 export type { IndexedElement, ValidationCode, ValidationIssue, ValidationSeverity } from './types';
 export { validateBundle, validateResource } from './utils';